Add view model to WriteSwatchView
