+++ FilamentSwatches/Views/_Components/NFCButtons/WriteNFCButton.swift
@@ -1,26 +1,25 @@
 //
-//  ReadNFCButton.swift
+//  WriteNFCButton.swift
 //  FilamentSwatches
 //
-//  Created by Jonas Frey on 07.11.22.
+//  Created by Jonas Frey on 10.11.22.
 //
 
-import AppFoundation
 import CoreNFC
 import DependencyInjection
 import Logging
 import SwiftUI
 
-struct ReadNFCButton: StatefulView {
+struct WriteNFCButton: View {
     @State var viewModel: ViewModel
     
     var body: some View {
         Button {
             Task {
-                await viewModel.readNFC()
+                await viewModel.writeNFC()
             }
         } label: {
-            Text("Read Tag")
+            Text("Write Tag")
                 .prominentButtonStyle()
         }
         .padding()
@@ -30,19 +29,16 @@ struct ReadNFCButton: StatefulView {
             Text("This device doesn't support tag scanning.")
         }
         .alert(
-            "Scanning Error",
+            "Writing Error",
             isPresented: $viewModel.isShowingGeneralError,
             actions: {
                 okayButton
             },
             message: {
                 let errorMessage = viewModel.generalError?.localizedDescription ?? String(localized: "An unknown error occurred.")
-                Text("There was an error scanning the NFC tag: \(errorMessage)")
+                Text("There was an error writing the NFC tag: \(errorMessage)")
             }
         )
-        .sheet(item: $viewModel.presentedSwatch) { swatch in
-            SwatchView(viewModel: .init(swatch: swatch))
-        }
     }
     
     private var okayButton: some View {
@@ -50,8 +46,8 @@ struct ReadNFCButton: StatefulView {
     }
 }
 
-struct ReadNFCButton_Previews: PreviewProvider {
+struct WriteNFCButton_Previews: PreviewProvider {
     static var previews: some View {
-        ReadNFCButton(viewModel: .init())
+        WriteNFCButton(viewModel: .init(swatch: SampleData.swatch))
     }
 }
